Character core v3
- Continued to split the body part classes.
- Removed a duplicated string variable.
- The character and the body part scripts will now be stored in separate folders.
