+++ Character-Generator/Assets/Scripts/Body Part.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e445bd9269ecb3a4993198b98db9cbb2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
