+++ Character-Generator/Assets/Scripts/Body Part/BodyPart.cs
@@ -1,0 +1,22 @@
+using UnityEngine;
+
+[System.Serializable]
+public class BodyPart {
+
+    public enum Part {
+
+        Head,
+        Torso,
+
+    };
+
+    [Space(6)]
+    public Part part;
+
+    [Space(6)]
+    public string customName;
+
+    [Space(6)]
+    public GameObject prefab;
+
+}
