+++ Character-Generator/Assets/Scripts/Body Part/BodyPart.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5f8069a820e54f94d8f3bcaa5c26e404
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
