+++ Character-Generator/Assets/Scripts/Body Part/CharacterBodyPart.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+
+[System.Serializable]
+public class CharacterBodyPart : BodyPart {
+
+    [Space(6)]
+    public Vector3 position;
+
+    [Space(6)]
+    public bool required;
+
+}
