+++ Character-Generator/Assets/Scripts/Body Part/CharacterBodyPart.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 33336c96050e00f489caa39ffe3c17be
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
